fix(i18n): IntoFluentValue<u16> + drop u16 cast in main.rs
CI's clippy turned up one more cast_lossless: `config.gateway.port as
i64` (port is u16). Adding `IntoFluentValue` for u16 so call sites pass
the value through unchanged, then dropping the explicit cast.

+++ src/main.rs
@@ -1108,7 +1108,7 @@ async fn main() -> Result<()> {
             println!("  {}", construct::t!("next-step-pairing-enabled"));
             println!(
                 "  {}",
-                construct::t!("next-step-dashboard", port = config.gateway.port as i64)
+                construct::t!("next-step-dashboard", port = config.gateway.port)
             );
             println!();
         }
